app/requirements.txt
Flask==2.0.1
psycopg2-binary==2.9.1
redis==4.1.0
requests==2.26.0
prometheus-client==0.11.0
python-dotenv==0.19.0
alembic==1.7.1
SQLAlchemy==1.4.23 